ALERT_rev.4

+++ src/main/java/com/alert/app/backend/api/gios/dto/GiosApiPollutionValueDto.java
@@ -16,5 +16,5 @@ public class GiosApiPollutionValueDto {
 
     @JsonFormat(shape=JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime date;
-    private long value;
+    private double value;
 }
